+++ src/auth/auth.controller.ts
@@ -11,14 +11,14 @@ import {
 } from '@nestjs/common';
 import { AuthService } from './auth.service';
 import { AuthGuard } from './auth.guard';
-import { JwtTokenPayload } from './jwt.token.payload';
+import { JwtTokenPayloadDto } from './dto/jwt.token.payload.dto';
 import { LoginRequestDto } from './dto/auth.request.dto';
 import { LoginResponseDto } from './dto/auth.response.dto';
 import { ApiBearerAuth } from '@nestjs/swagger';
 import { User } from '../common/user.decorator';
 import { UserService } from '../user/user.service';
-import { UserResultDto } from '../user/user.response.dto';
-import { SignupRequestDto } from '../user/user.request.dto';
+import { UserResultDto } from '../user/dto/user.response.dto';
+import { SignupRequestDto } from '../user/dto/user.request.dto';
 
 @Controller('auth')
 export class AuthController {
@@ -44,11 +44,11 @@ export class AuthController {
 
   // 인증 로직 동작 확인용 API
   @UseInterceptors(ClassSerializerInterceptor)
-  @SerializeOptions({ type: JwtTokenPayload })
+  @SerializeOptions({ type: JwtTokenPayloadDto })
   @ApiBearerAuth()
   @UseGuards(AuthGuard)
   @Get('/info')
-  getTokenData(@User() user: JwtTokenPayload): JwtTokenPayload {
+  getTokenData(@User() user: JwtTokenPayloadDto): JwtTokenPayloadDto {
     //이 API를 실제로 사용해서는 안 됨
     if (process.env.NODE_ENV === 'prod') {
       throw new ForbiddenException();
