+++ src/auth/auth.guard.ts
@@ -5,9 +5,8 @@ import {
   UnauthorizedException,
   UnsupportedMediaTypeException,
 } from '@nestjs/common';
-import { Request } from 'express';
 import { JwtService } from '@nestjs/jwt';
-import { JwtTokenPayload } from './jwt.token.payload';
+import { JwtTokenPayloadDto } from './dto/jwt.token.payload.dto';
 import { GqlExecutionContext } from '@nestjs/graphql';
 
 @Injectable()
@@ -17,7 +16,7 @@ export class AuthGuard implements CanActivate {
   async canActivate(
     context: ExecutionContext & { contextType: string },
   ): Promise<boolean> {
-    let request: Request;
+    let request: { headers: { authorization: string } };
     switch (context.contextType) {
       case 'http':
         request = context.switchToHttp().getRequest();
@@ -26,7 +25,7 @@ export class AuthGuard implements CanActivate {
       case 'graphql':
         // 이 부분은 Type 단언을 하지 않으면 getContext()의 결과값이 any 타입으로 취급되어 ESlint 경고가 발생합니다.
         // 이에 따라 코드를 작성했는데, 다른 방식(getContext의 type을 별도로 확인하는 코드 작성)을 희망하시면 이에 따라 수정하겠습니다.
-        request = <Request>GqlExecutionContext.create(context).getContext<{
+        request = GqlExecutionContext.create(context).getContext<{
           req: { headers: { authorization: string } };
         }>().req;
         break;
@@ -42,7 +41,7 @@ export class AuthGuard implements CanActivate {
     }
 
     try {
-      const payload: JwtTokenPayload = await this.jwtService.verifyAsync(
+      const payload: JwtTokenPayloadDto = await this.jwtService.verifyAsync(
         token,
         {
           secret: process.env.JWT_SECRET,
