+++ src/auth/dto/jwt.token.payload.dto.ts
@@ -1,7 +1,7 @@
 import { Exclude, Expose, Type } from 'class-transformer';
 
 @Exclude()
-export class JwtTokenPayload {
+export class JwtTokenPayloadDto {
   @Expose()
   @Type(() => BigInt)
   id: bigint;
