+++ src/user/dto/user.response.dto.ts
@@ -2,6 +2,7 @@ import { Exclude } from 'class-transformer';
 
 export class UserResultDto {
   id: bigint;
+
   loginId: string;
 
   @Exclude() // return 시 보이지 않게 설정
