+++ src/user/user.resolver.ts
@@ -6,7 +6,7 @@ import {
   ResolveField,
   Resolver,
 } from '@nestjs/graphql';
-import { UserDto } from './user.dto';
+import { UserDto } from './dto/user.dto';
 import { UserService } from './user.service';
 
 @Resolver(() => UserDto)
