+++ src/vote/vote.enum.ts
@@ -6,7 +6,6 @@ export enum VoteStatus {
   ENDED = 'ENDED',
 }
 
-// TODO: register 위치가 적절한지 확인 필요
 registerEnumType(VoteStatus, {
   name: 'VoteStatus',
 });
